pyproject.toml
[project] 
name = "decision-intelligence-env" 
version = "1.0.0" 
requires-python = ">=3.10" 
dependencies = [ 
    "fastapi", 
    "uvicorn", 
    "numpy", 
    "openenv-core>=0.2.0", 
] 
 
[project.scripts]
openenv-server = "server.app:main"
